+++ src/host/pc/time.c
@@ -31,7 +31,7 @@ char *P_ctime(Uint t)
 	    (buf[4] != 'F' || (buf[8] == '2' && buf[9] == '9'))) {
 	    /* 2100 is not a leap year */
 	    t += 86400;
-	    if (t < 0) {
+	    if ((Int) t < 0) {
 		t -= 1009843200;
 		offset += 32;
 	    }
