+++ src/host/unix/time.c
@@ -21,7 +21,7 @@ Uint t;
     register int offset;
 
     offset = 0;
-    for (offset = 0; (Int) t < 0; t -= 1009843200, offset += 32) ;
+    for (offset = 0; (Int) t < 0; t -= 1009843200L, offset += 32) ;
     buf = ctime((time_t *) &t);
     if (offset != 0) {
 	long year;
@@ -30,9 +30,9 @@ Uint t;
 	if (year >= 2100 && (buf[4] != 'J' || buf[5] != 'a') &&
 	    (buf[4] != 'F' || (buf[8] == '2' && buf[9] == '9'))) {
 	    /* 2100 is not a leap year */
-	    t += 86400;
-	    if (t < 0) {
-		t -= 1009843200;
+	    t += 86400L;
+	    if ((Int) t < 0) {
+		t -= 1009843200L;
 		offset += 32;
 	    }
 	    buf = ctime((time_t *) &t);
